% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handler.R
\name{get_r_data_by_year}
\alias{get_r_data_by_year}
\title{Get data from GitHub repo}
\usage{
get_r_data_by_year(year)
}
\arguments{
\item{year}{}
}
\value{
\code{tibble()}
}
\description{
Get data from GitHub repo
}
\examples{
\dontrun{
get_r_data_by_year(2000)
}
}
