% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{info_repo}
\alias{info_repo}
\title{Informations about data file}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 75 rows and 5 columns.
}
\usage{
info_repo
}
\description{
Formats, HASHs, sizes
}
\keyword{datasets}
