% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meteobr.max.year}
\alias{meteobr.max.year}
\title{Package maximum yaer available}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
meteobr.max.year
}
\description{
Package maximum yaer available
}
\keyword{datasets}
