% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handler.R
\name{get_data}
\alias{get_data}
\title{Get data between dates}
\usage{
get_data(first.day, last.day, vars = NULL, stations = NULL)
}
\arguments{
\item{first.day}{String like "mm-dd".}

\item{last.day}{String like "mm-dd".}

\item{vars}{(optional) Variables to be collected.}

\item{stations}{(optional) Stations to be collected.}
}
\value{
A tibble.
}
\description{
\code{get_data()} does this.
Beware that this function, check \code{local_data()} to see more.
}
