% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handler.R
\name{get_inmet_data_by_year}
\alias{get_inmet_data_by_year}
\title{Title}
\usage{
get_inmet_data_by_year(
  year,
  first.day = NA,
  last.day = NA,
  vars = NULL,
  stations = NULL,
  local.path = local_data()
)
}
\arguments{
\item{local.path}{}
}
\description{
Title
}
