% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{info_repo}
\alias{info_repo}
\title{Some info about data in github repo}
\format{
\subsection{\code{info_repo}}{

A data frame with 75 rows and 5 columns:
\describe{
\item{year}{file's year}
\item{type}{type: either Rdata, zip, fst}
\item{file}{full file name}
\item{size}{size in Mb}
\item{hash}{md5 hash}
}
}
}
\usage{
info_repo
}
\description{
Some info about data in github repo
}
\keyword{datasets}
