% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handler.R
\name{set_data_locally}
\alias{set_data_locally}
\title{Download preprocessed data}
\usage{
set_data_locally(years = 2000:meteobr.max.year)
}
\arguments{
\item{years}{Vector of integers between 2000 and 2024; if no value is passed,
the DEFAULT is to download all data.}
}
\description{
\code{set_data_locally()} downloads .Rdata files and stores it locally without the
hassle of unzipping and processing them.
}
\examples{
\dontrun{
# Should download all data available on
# <https://github.com/carlosdemoura/meteobr/raw/refs/heads/master/data/repo/>
# and store it where `local_data()` says.
set_data_locally()
}

}
