% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stations}
\alias{stations}
\title{Additional information about INMET's automatic stations}
\format{
\subsection{\code{stations}}{

A data frame with 624 rows and 9 columns:
\describe{
\item{station.id}{INMET's id of the station}
\item{station.name}{original name of the station on INMET's website}
\item{town.id}{IBGE code for station's town}
\item{town.name}{name of the town in which station is located, standardized}
\item{state}{state in which the station is located (abbreviation)}
\item{region}{national region in which the station is located}
\item{lat}{latitude of the station (meters)}
\item{lon}{longitude of the station (meters)}
\item{alt}{altitude of the station (meters)}
}
}
}
\usage{
stations
}
\description{
Additional information about INMET's automatic stations
}
\keyword{datasets}
